[1.20.4-1.20.6] v1.1.0 - Stable
Fixed hopefully all the bugs from the previous beta version. Check out the changelog.md for all the changes in this version!

+++ common/src/main/java/dev/mrsnowy/teleport_commands/storage/StorageManager.java
@@ -142,9 +142,9 @@ public class StorageManager {
 
             public static class Home {
                 public String name;
-                public double x;
-                public double y;
-                public double z;
+                public int x;
+                public int y;
+                public int z;
                 public String world;
             }
         }
